Refactor legacy app
