+++ src/app/LegacyTradeApp/pages/About.stories.tsx
@@ -1,11 +1,11 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0'
 
-import FAQ from 'pages/FAQ'
+import About from './About'
 
 export default {
-  title: 'Pages/FAQ',
-  component: FAQ,
+  title: 'Pages/About',
+  component: About,
 } as Meta
 
-export const Primary = FAQ
+export const Primary = About
