+++ src/app/LegacyTradeApp/pages/About.tsx
