+++ src/app/LegacyTradeApp/pages/ConnectWallet.tsx
