+++ src/app/LegacyTradeApp/pages/FAQ.stories.tsx
@@ -1,11 +1,11 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta } from '@storybook/react/types-6-0'
 
-import About from 'pages/About'
+import FAQ from './FAQ'
 
 export default {
-  title: 'Pages/About',
-  component: About,
+  title: 'Pages/FAQ',
+  component: FAQ,
 } as Meta
 
-export const Primary = About
+export const Primary = FAQ
