+++ src/app/LegacyTradeApp/pages/FAQ.tsx
