+++ src/app/LegacyTradeApp/pages/Strategies.tsx
