+++ src/app/LegacyTradeApp/pages/Trade.tsx
