+++ src/app/LegacyTradeApp/pages/Trades.tsx
